+++ src/identity/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @module Identity
+ * @experimental
+ */
+
+export * from './derivation.js';
+export * from './get-private-key.js';
+export * from './identity.js';